Fix up graph_for
